+++ packages/widget/src/constants/bridgeable.ts
@@ -609,6 +609,7 @@ export const SYN: BridgeableToken = {
     [CHAINS.DOGE.id]: '0xDfA53EeBA61D69E1D2b56b36d78449368F0265c1',
     [CHAINS.CANTO.id]: '0x555982d2E211745b96736665e19D9308B615F78e',
     [CHAINS.BASE.id]: '0x432036208d2717394d2614d6697c46DF3Ed69540',
+    [CHAINS.BLAST.id]: '0x9592f08387134e218327E6E8423400eb845EdE0E',
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 18,
@@ -628,6 +629,7 @@ export const SYN: BridgeableToken = {
     [CHAINS.DOGE.id]: 18,
     [CHAINS.CANTO.id]: 18,
     [CHAINS.BASE.id]: 18,
+    [CHAINS.BLAST.id]: 18,
   },
   symbol: 'SYN',
   name: 'Synapse',
@@ -697,6 +699,7 @@ export const NUSD: BridgeableToken = {
     [CHAINS.AVALANCHE.id]: '0xCFc37A6AB183dd4aED08C204D1c2773c0b1BDf46',
     [CHAINS.HARMONY.id]: '0xED2a7edd7413021d440b09D654f3b87712abAB66',
     [CHAINS.AURORA.id]: '0x07379565cD8B0CaE7c60Dc78e7f601b34AF2A21c',
+    [CHAINS.BLAST.id]: '0x3194B0A295D87fDAA54DF852c248F7a6BAF6c6e0',
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 18,
@@ -712,6 +715,7 @@ export const NUSD: BridgeableToken = {
     [CHAINS.AVALANCHE.id]: 18,
     [CHAINS.HARMONY.id]: 18,
     [CHAINS.AURORA.id]: 18,
+    [CHAINS.BLAST.id]: 18,
   },
   symbol: 'nUSD',
   name: 'Synapse nUSD',
@@ -751,6 +755,7 @@ export const NETH: BridgeableToken = {
     [CHAINS.ARBITRUM.id]: '0x3ea9B0ab55F34Fb188824Ee288CeaEfC63cf908e',
     [CHAINS.AVALANCHE.id]: '0x19E1ae0eE35c0404f835521146206595d37981ae',
     [CHAINS.HARMONY.id]: '0x0b5740c6b4a97f90eF2F0220651Cca420B868FfB',
+    [CHAINS.BLAST.id]: '0xce971282fAAc9faBcF121944956da7142cccC855',
   },
   decimals: {
     [CHAINS.OPTIMISM.id]: 18,
@@ -762,6 +767,7 @@ export const NETH: BridgeableToken = {
     [CHAINS.ARBITRUM.id]: 18,
     [CHAINS.AVALANCHE.id]: 18,
     [CHAINS.HARMONY.id]: 18,
+    [CHAINS.BLAST.id]: 18,
   },
   symbol: 'nETH',
   name: 'Synapse nETH',
@@ -782,6 +788,7 @@ export const ETH: BridgeableToken = {
     [CHAINS.BASE.id]: ZeroAddress,
     [CHAINS.ARBITRUM.id]: ZeroAddress,
     [CHAINS.DFK.id]: '0xfBDF0E31808d0aa7b9509AA6aBC9754E48C58852',
+    [CHAINS.BLAST.id]: ZeroAddress,
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 18,
@@ -791,6 +798,7 @@ export const ETH: BridgeableToken = {
     [CHAINS.BASE.id]: 18,
     [CHAINS.ARBITRUM.id]: 18,
     [CHAINS.DFK.id]: 18,
+    [CHAINS.BLAST.id]: 18,
   },
   symbol: 'ETH',
   name: 'Ethereum',
@@ -1237,6 +1245,7 @@ export const WETH: BridgeableToken = {
     [CHAINS.KLAYTN.id]: '0xCD6f29dC9Ca217d0973d3D21bF58eDd3CA871a86',
     [CHAINS.BASE.id]: '0x4200000000000000000000000000000000000006',
     [CHAINS.ARBITRUM.id]: '0x82af49447d8a07e3bd95bd0d56f35241523fbab1',
+    [CHAINS.BLAST.id]: '0x4300000000000000000000000000000000000004',
   },
   decimals: {
     [CHAINS.ETHEREUM.id]: 18,
@@ -1248,6 +1257,7 @@ export const WETH: BridgeableToken = {
     [CHAINS.KLAYTN.id]: 18,
     [CHAINS.BASE.id]: 18,
     [CHAINS.ARBITRUM.id]: 18,
+    [CHAINS.BLAST.id]: 18,
   },
   symbol: 'WETH',
   name: 'Wrapped ETH',
@@ -1311,4 +1321,21 @@ export const USDBC: BridgeableToken = {
   routeSymbol: 'USDbC',
   imgUrl:
     'https://www.synapseprotocol.com/_next/static/media/usdc.f9b139bb.svg',
+}
+
+export const USDB: BridgeableToken = {
+  addresses: {
+    [CHAINS.BLAST.id]: '0x4300000000000000000000000000000000000003',
+  },
+  decimals: {
+    [CHAINS.BLAST.id]: 18,
+  },
+  symbol: 'USDB',
+  name: 'Blast Rebasing USD',
+  swapableType: 'USD',
+  color: 'yellow',
+  priorityRank: 125,
+  routeSymbol: 'USDB',
+  imgUrl:
+    'https://www.synapseprotocol.com/_next/static/media/usdb.33190604.svg',
 }