+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.26](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.25...@synapsecns/sdk-router@0.3.26) (2024-03-01)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.3.25](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.3.24...@synapsecns/sdk-router@0.3.25) (2024-02-22)
 
 **Note:** Version bump only for package @synapsecns/sdk-router